requirements.txt
fastapi
uvicorn
python-dotenv
motor
requests
beanie
passlib
jinja2
bcrypt
pillow
fastapi-mail
pytest
pytest-asyncio